fix: RateRepository 구현체 타입을 매개변수로 받아서 처리하기
