+++ src/data/repositories/RateRepository.ts
@@ -1,0 +1,26 @@
+export interface Rate {
+  id: number;
+  rate: number;
+}
+
+interface RateRepository {
+  getMovieRate(id: number): number;
+  setMovieRate(id: number, rate: number): void
+}
+
+export class RateRepositoryImpl implements RateRepository {
+  #getRates(){
+    const rates = localStorage.getItem('rates') || "{}";
+    return JSON.parse(rates) || {};
+  }
+  getMovieRate(id: number){
+    const rates = this.#getRates();
+
+    return rates[id];
+  }
+  setMovieRate(id: number, rate: number){
+    const prevRates = this.#getRates();
+    const rates = { ...prevRates, [id]: rate };
+    localStorage.setItem('rates', JSON.stringify(rates));
+  }
+}
